Split cql command func into internal package
