+++ cmd/cql/internal/adapter.go
@@ -1,0 +1,1 @@
+package internal
