+++ cmd/cql/internal/base.go
@@ -1,0 +1,1 @@
+package internal
