+++ cmd/cql/internal/command.go
@@ -1,0 +1,184 @@
+/*
+ * Copyright 2016-2018 Kenneth Shaw.
+ * Copyright 2018 The CovenantSQL Authors.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package internal
+
+import (
+	"context"
+	"database/sql"
+	"database/sql/driver"
+	"errors"
+	"fmt"
+	"reflect"
+	"strconv"
+	"strings"
+	"time"
+
+	sqlite3 "github.com/CovenantSQL/go-sqlite3-encrypt"
+	"github.com/sirupsen/logrus"
+	"github.com/xo/dburl"
+	"github.com/xo/usql/drivers"
+	"github.com/xo/usql/text"
+
+	"github.com/CovenantSQL/CovenantSQL/client"
+)
+
+// SqTime provides a type that will correctly scan the various timestamps
+// values stored by the github.com/mattn/go-sqlite3 driver for time.Time
+// values, as well as correctly satisfying the sql/driver/Valuer interface.
+type SqTime struct {
+	time.Time
+}
+
+// Value satisfies the Valuer interface.
+func (t SqTime) Value() (driver.Value, error) {
+	return t.Time, nil
+}
+
+// Scan satisfies the Scanner interface.
+func (t *SqTime) Scan(v interface{}) error {
+	switch x := v.(type) {
+	case time.Time:
+		t.Time = x
+		return nil
+	case []byte:
+		return t.parse(string(x))
+	case string:
+		return t.parse(x)
+	}
+
+	return fmt.Errorf("cannot convert type %s to time.Time", reflect.TypeOf(v))
+}
+
+// parse attempts to parse string s to t.
+func (t *SqTime) parse(s string) error {
+	if s == "" {
+		return nil
+	}
+
+	for _, f := range sqlite3.SQLiteTimestampFormats {
+		z, err := time.Parse(f, s)
+		if err == nil {
+			t.Time = z
+			return nil
+		}
+	}
+
+	return errors.New("could not parse time")
+}
+
+// UsqlRegister init xo/usql driver
+func UsqlRegister(log *logrus.Logger, waitTxConfirmationMaxDuration time.Duration, dsn string) {
+	// set command name of usql
+	text.CommandName = "covenantsql"
+
+	// register SQLite3 database
+	drivers.Register("sqlite3", drivers.Driver{
+		AllowMultilineComments: true,
+		ForceParams: drivers.ForceQueryParameters([]string{
+			"loc", "auto",
+		}),
+		Version: func(db drivers.DB) (string, error) {
+			var ver string
+			err := db.QueryRow(`SELECT sqlite_version()`).Scan(&ver)
+			if err != nil {
+				return "", err
+			}
+			return "SQLite3 " + ver, nil
+		},
+		Err: func(err error) (string, string) {
+			if e, ok := err.(sqlite3.Error); ok {
+				return strconv.Itoa(int(e.Code)), e.Error()
+			}
+
+			code, msg := "", err.Error()
+			if e, ok := err.(sqlite3.ErrNo); ok {
+				code = strconv.Itoa(int(e))
+			}
+
+			return code, msg
+		},
+		ConvertBytes: func(buf []byte, tfmt string) (string, error) {
+			// attempt to convert buf if it matches a time format, and if it
+			// does, then return a formatted time string.
+			s := string(buf)
+			if s != "" && strings.TrimSpace(s) != "" {
+				t := new(SqTime)
+				if err := t.Scan(buf); err == nil {
+					return t.Format(tfmt), nil
+				}
+			}
+			return s, nil
+		},
+	})
+
+	// register CovenantSQL database
+	drivers.Register("covenantsql", drivers.Driver{
+		AllowMultilineComments: true,
+		Version: func(db drivers.DB) (string, error) {
+			return Version, nil
+		},
+		Err: func(err error) (string, string) {
+			return "", err.Error()
+		},
+		ConvertBytes: func(buf []byte, tfmt string) (string, error) {
+			// attempt to convert buf if it matches a time format, and if it
+			// does, then return a formatted time string.
+			s := string(buf)
+			if s != "" && strings.TrimSpace(s) != "" {
+				t := new(SqTime)
+				if err := t.Scan(buf); err == nil {
+					return t.Format(tfmt), nil
+				}
+			}
+			return s, nil
+		},
+		RowsAffected: func(sql.Result) (int64, error) {
+			return 0, nil
+		},
+		Open: func(url *dburl.URL) (handler func(driverName, dataSourceName string) (*sql.DB, error), err error) {
+			log.Infof("connecting to %#v", url.DSN)
+
+			// wait for database to become ready
+			ctx, cancel := context.WithTimeout(context.Background(), waitTxConfirmationMaxDuration)
+			defer cancel()
+			if err = client.WaitDBCreation(ctx, dsn); err != nil {
+				return
+			}
+
+			return sql.Open, nil
+		},
+	})
+
+	// register covenantsql:// scheme to dburl
+	dburl.Register(dburl.Scheme{
+		Driver: "covenantsql",
+		Generator: func(url *dburl.URL) (string, error) {
+			dbID, err := dburl.GenOpaque(url)
+			if err != nil {
+				return "", err
+			}
+			cfg := client.NewConfig()
+			cfg.DatabaseID = dbID
+			return cfg.FormatDSN(), nil
+		},
+		Proto:    0,
+		Opaque:   true,
+		Aliases:  []string{},
+		Override: "",
+	})
+}
