+++ cmd/cql/internal/drop.go
@@ -1,0 +1,1 @@
+package internal
