+++ cmd/cql/internal/help.go
@@ -1,0 +1,9 @@
+package internal
+
+// Name of command
+const Name = "cql"
+
+var (
+	// Version of command, set by main func of version
+	Version = "unknown"
+)
