+++ cmd/cql/internal/transfer.go
@@ -1,0 +1,1 @@
+package internal
