+++ cmd/cql/internal/web.go
@@ -1,0 +1,1 @@
+package internal
